Refactored validation.ts and index.ts (split into files)
